requirements.txt
# All dependencies are installed directly in Dockerfile
# See Dockerfile for the complete list of dependencies and versions 

torch==1.13.1
transformers==4.30.2
nltk==3.8.1
python-dotenv==1.0.0
requests==2.31.0
boto3>=1.26.0
beautifulsoup4==4.12.3 